332853: Fix some FindBugs warnings - use of file.encoding when acquiring dumps https://bugs.eclipse.org/bugs/show_bug.cgi?id=332853

+++ plugins/org.eclipse.mat.hprof/src/org/eclipse/mat/hprof/Pass1Parser.java
@@ -568,7 +568,7 @@ public class Pass1Parser extends AbstractParser
 
         in.skipBytes((long) size * idSize);
         previousArrayStart = address;
-        previousArrayUncompressedEnd = address + 16 + size * 8;
+        previousArrayUncompressedEnd = address + 16 + (long)size * 8;
     }
 
     private void readPrimitiveArrayDump(long segmentStartPos) throws SnapshotException, IOException
